requirements.txt
Flask
openai
PyMuPDF
python-docx
